Fixed circ dependency in cli package
