+++ python/src/aac/__main__.py
@@ -7,6 +7,6 @@ if sys.argv[0].endswith("__main__.py"):
 
 
 if __name__ == "__main__":
-    from aac import cli
+    from aac.cli.execute import run_cli
 
-    cli.run_cli()
+    run_cli()
