+++ python/src/aac/plugins/start_lsp/start_lsp_impl.py
@@ -2,34 +2,14 @@
 # NOTE: It is safe to edit this file.
 # This file is only initially generated by the aac gen-plugin, and it won't be overwritten if the file already exists.
 
-import logging
-
-from typing import Optional
-
-from pygls.server import LanguageServer
-
-from aac.lang.server import setup_features
+import aac.lang.lsp.server as lsp_server
 from aac.plugins.plugin_execution import PluginExecutionResult, plugin_result
 
 plugin_name = "start-lsp"
 
-server: Optional[LanguageServer] = None
-
 
 def start_lsp() -> PluginExecutionResult:
     """Start the IO LSP server."""
-
-    def start():
-        global server
-
-        server = server or LanguageServer()
-        setup_features(server)
-
-        logger = logging.getLogger(__name__)
-        logger.info("Starting the AaC LSP server")
-
-        server.start_io()
-
-    with plugin_result(plugin_name, start) as result:
+    with plugin_result(plugin_name, lsp_server.start_lsp) as result:
         result.messages = [m for m in result.messages if m]
         return result
