+++ python/tests/lang/test_server.py
@@ -42,17 +42,5 @@ class TestLspServer(TestCase):
         self.assertSequenceEqual(list(res.keys()), ["contents"])
         self.assertIn("LSP server", res.get("contents"))
 
-    def test_handles_completion_request(self):
-        active_context = get_active_context(reload_context=True)
-
-        res: list[CompletionItem] = self.client.send_request(
-            methods.COMPLETION,
-            CompletionParams(text_document={"uri": TEST_DOCUMENT_URI}, position=Position(line=0, character=0)),
-        )
-
-        self.assertSequenceEqual(list(res.keys()), ["isIncomplete", "items"])
-        self.assertSequenceEqual([i.get("label") for i in res.get("items")], active_context.get_root_keys())
-        self.assertFalse(res.get("isIncomplete"))
-
 
 TEST_DOCUMENT_URI = uris.from_fs_path(__file__)
